backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.1
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.0
sentence-transformers==2.2.2 